--- next.config.mjs
@@ -1,16 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {};
-
-const withPWA = require("@ducanh2912/next-pwa").default({
-    dest: "public",
-    cacheOnFrontEndNav: false,
-    aggressiveFrontEndNavCaching: true,
-    reloadOnOnline: true,
-    swcMinify: true,
-    disable: process.env.NODE_ENV === "development",
-    workboxOptions: {
-        disableDevLogs: true,
-    },
-});
-
-export default withPwa(nextConfig)
